Add vi mode indicator to start of zsh prompt
I use vi mode within zsh, and stumbled onto a way to add the mode
indicator to the prompt this evening.  It should address one of my
pain points with editing a line with vi mode, when it doesn't
quite act as I expect.  Makes it easier to tell there's a delay
when switching to Normal mode.