Add vim expansion shortcut for current buffer directory
In vimrd, define mapping of '%%' to '%:h<Tab' when in command
mode.  This allows `:edit %%` to expand to `:edit <current buffer's path>`,
which can be handy for opening files in the same directory.

This config taken from 'Practical Vim' by Drew Neil.  (Working my way
through that book now, already learned enough from it to make it worth
the purchase.)