Change zsh prompt vi mode indicators
The zsh configuration enables vi mode, and the prompt begins with
an indicator of whether vi's insert or normal mode is active.
Previously the indicator was a cyan "-" for insert and a magenta
"+" for normal mode.  This changes insert mode to a curved arrow
character, and normal mode is a star character.

Aside from being a bit more interesting to look at, I've found it
also makes prompt lines stand out more from other console output.
Trying it out for now to see how I like it longer term.