Bind ctrl-c in tmux to create and rename window
In tmux, c is bound to create a window.  Usually when I create a
window, I immediately rename it.  With this binding, instead
of `<prefix> c <prefix> ,` to create a window and rename, I can
just do `<prefix> ctrl-c`.

My first choice for this binding would have been shift-c, but I
wasn't able to get that to work.