Add tmdir script
'tmdir' is a utility script that creates a tmux session containing
a window for each subdirectory of the current directory.  It includes
options for opening a horizontal or vertical split in each window, as
well as an option to open your $EDITOR in the first split of each
window.

For example:
tmdir -he -n myproject

This will create a tmux session named 'myproject'.  For each
subdirectory of the current directory, a horizontal split will be
created, and vim will be open in the left split of each window.

Expecting this will save me a bit of time at work with tmux workspace
setup.  (I should get better with tmux-resurrection and similar plugins
too.)