Extend tmdir script to support config file
The tmdir script creates a tmux session with windows for the
subdirectories of the current directory.  This is useful and simple,
but there is no way to limit the subdirectory windows opened, and there
is no way to control the names for windows when they are created.

In this change, if the tmdir command is run in a directory that contains
a .tmdir file, the file will be interpreted as a list of subdirectories
for which a window will be created.  Also, if the directory name is
followed by a colon and a string, the string will be used as the tmux
window name.