Add vim-instant-markdown plugin
Vim plugin that live-updates a browser window with HTML-formatted
Markdown when editing .md files.  Kinda spiffy, I used to use a separate
Mac app for that kind of thing.  I don't want that popping up every
single time I edit a Markdown file though, so I bound it to <leader>md.

For this plugin to work, the instant-markdown-d npm module needs to be
installed globally.  This commit includes an update to
packages/npm-g-list as a result.