Add vim-highlightedyank plugin
Vim plugin to briefly highlight the text just yanked.  Only drawback I
see is it doesn't work with 'Y' to highlight a full line.. but 'yy' also
highlights a full line, so it should be easy enough for me to switch.
Forgot about that version, will see which I end up liking better.