Add Syntastic to vimrc, update statusline
Enable Syntastic plugin for syntax validation.  I had this configured
but disabled for some time, and now that I'm getting more used to it,
it's time to turn it back on.

Also the statusline changes might still need work, but I think it's an
improvement in some spots already - showing the filetype and encoding,
for example.  This statusline includes the Syntastic indicators as well.