Add vim-mucomplete plugin for autocomplete
For now, giving up on getting autocomplete working in vim with
YouCompleteMe, and instead trying a different plugin I learned about
called vim-mucomplete.  This plugin is a lighter wrapper around
vim's built-in autocomplete.  So far I like it; will try it for a while
and see if it sticks for me.