Change move-lines vim mapping from ctrl- to alt-
One aspect of my workflow that I've been meaning to improve upon is
the context switching between tmux and vim pane navigation.  I read a
blog post today that describes a sensible-sounding setup using ctrl-j
and ctrl-k to seamlessly move between both pane types... which is great
except I'm already using those keybindings to move lines up and down
(and I use that pretty regularly).

Solution is to switch the move-line keybindings to alt-j and alt-k.
There is a quirk to that on Macs though, we have to use the actual
character sent to the terminal instead of "<Alt-j>" because that's what
gets sent through the terminal to Vim.

Not tested on gvim (which I don't really use), but it does work for me
within tmux.