Enable tmux/vim pane switching with ctrl-h/j/k/l
Add vim-tmux-navigator plugin to vimrc, and add equivalent tmux plugin.

I ran into an issue with zsh: ctrl-h was triggering the 'delete word'
action instead of switching tmux pane, so I ended up defining a custom
zle widget that overrides the ctrl-h mapping when in tmux.  (Learned
that a special key sequence needs to be sent by iTerm2 for this to work
well too.)

Finally, I adjusted a few other mappings related to this change.  I
removed the ctrl-j/k mappings for Ultisnips for now -- will need to find
an alternate, but I don't use snippets nearly as much as I switch panes.
I added a new tmux mapping to support refreshing the vim display since
the usual ctrl-l won't be available for that task either.