Swallow stderr in 'tm' zsh function
The 'tm' function is defined as a shortcut for either creating or
attaching to a tmux session; with autocomplete it's much handier than
remembering or typing out the equivalent tmux commands.  But as defined,
if tmux isn't already running, starting a new session would print an
error like "no server running on /private/tmp/tmux-501/default" before
opening the tmux session.  Annoying.  This change gets rid of that
error.