Add antibody, autocomplete, and syntax color to zsh
Including antibody to load zsh plugins.  Plugins are added via a
separate file in zsh directory.

Not 100% sure I like the autocomplete and syntax coloring yet, but
it looks interesting enough that I'm willing to give it a try.
Including antibody should make it easy to try out additional plugins
from oh-my-zsh and elsewhere if I find myself inspired to do so.