Alias 'cd ..' to dash in zshrc
Similar to using dash in vim's netrw to go up a directory level.

So far I haven't seen this alias cause any issues with other commands.
I'll be quick to pull this one if it does.  Maybe I should be working
on memorizing fasd shortcuts instead, but this shortcut has seemed
handy so far.