Add calculator function to zshrc
For those times where a calculator would be handy.  I sit with a command
prompt in front of me all day; I shouldn't need to switch to a
calculator widget to make the computer do some arithmetic for me! :)

Function is named "=".  Requires the expression to be surrounded by
quotes. Example:

  $ = "6*7"
  42
  $

Uses zsh's arithmetic expression evaluator.