Add todo.txt
Include todo-txt in Homebrew package list, Dropbox in Homebrew Cask
list, and update zsh and vim configs.

As fun as tweaking keybindings and vim plugins is, the next real area I
need to improve my general effectiveness is my organization skills and
habits.  To that end I'm going to give todo.txt combined with Dropbox a
try.

http://todotxt.com/ is a shell script/CLI for managing a todo list in a
plain text file.  There is an Android app for it that I'll probably buy
that works with Dropbox for syncing between devices, so I've joined the
modern era and set up a Dropbox account as well.