Update vim <leader>p binding in visual mode to preserve clipboard
When text is selected in visual select mode, using '<leader>p' will paste
over the selection without replacing the clipboard contents.  If for
some reason I wanted to replace the clipboard in that case, using just
'p' would accomplish that.  This makes replacing text with the clipboard
in more than one point in a file more convenient, and more consistent
with how other editors handle pasting.