Add vim plugins for textobjs, closetag, vim-eunuch
Discovered some additional text object plugins for VIm that will likely
come in handy.  Entire buffer text area will be handy for copying full
doc to clipboard, or for deleting a git commit buffer to cancel the
commit.  Copying a line of text without the trailing newline will also
be helpful.

(Need to keep an eye on whether these plugins are slowing down vim too
much though.)

vim-eunuch looks super helpful.  Automatically doing a chmod for scripts
that start with a hashbang is sweet.  Moving files and buffer in same
command is also going to be handy.

The functionality in closetag.vim has been on my to-learn list for a
while - easily close XML/HTML tags with a keystroke.