Update 'tm' tmux script to work within session
The tm script provides a shortcut for creating a tmux session if it
doesn't already exist, and attaching to a session if it does.  It's set
up with autocomplete for the session list as well.

However, to create or switch between sessions while within a session,
it's been necessary to detach from the session first.  This change
makes it so `tm` works within a session, allowing fast session switching
in addition to creating and attaching.