Update calendar.vim shortcut to better handle weekends
By default, when calendar.vim opens, the new panel is not wide enough
due to line numbers appearing in the left side.  Because of this, I have
my shortcut to open calendar.vim widen the panel.  This mostly works,
but on Saturdays the calendar is still partially scrolled off the
screen.  Adding '0' to my shortcut scrolls the calendar back into view,
and issuing the 't' command places the cursor back on today's date.