Remove ctrl-a as tmux command prefix
I had ctrl-a as a tmux command prefix when I first started learning and
getting used to tmux, but I haven't actually used it in ages.  It's only
a problem when I want to use ctrl-a within vim to increment digits.  But
I want to increment digits now and again, so time to fix the tmux
config.