Add vimrc-plugin-local support to vimrc
Allow vim plugins to be defined in local dotfiles within
$DOTFILE_LOCAL_HOME/vimrc-plugin-local file.  At work I have at least
one plugin that is specific to the company's internal tools.  By
allowing this plugin to be set within local dotfiles, I have one less
reason for my work dotfiles to be different from what's committed to the
public repo (and one less opportunity for merge conflicts when I try to
update).