Add 'cddocuments' alias
Similar to cddotfiles, cddesktop, and other aliases to cd to common
directories.

This change also moves $VIMWIKI_DIR to ~/Documents instead of the remote
sync directory.  Solution for making the vimwiki directory sync is to
create the real directory in the remote sync directory, and symlink it
to ~/Documents/vimwiki.