Add alias to cd to vimwiki diary directory
I've gotten in the habit of using the vimwiki diary every day at work,
and even most days at home as well.  Adding a 'cddiary' alias to go to
the diary directory to make it more convenient to search through my
daily notes. (It's not too terrible within vim already, but this makes it
a little nicer to use ripgrep or ag from the shell.)