Add personal-fortune to vimrc Startify header config
Update vimrc to show messages in $DOTFILE_LOCAL_HOME/personal-fortune
(if it exists).  Each message in the file should be separated by an
empty line.

The fortune message default in Startify is fun, but I thought it might
be more useful to have it display reminders or messages about long-term
goals or areas of improvement.

We recently had Annual Reviews at the mega-corp where I work.  The
feedback was good, but it would be more useful if I see the improvement
ideas regularly, rather than filing it away and not seeing it again
until next year (as usual).  I thought having them appear in my editor
startup screen might be a good place to have them pop up.