Switch vim plugin manager from Vundle to vim-plug
Vundle has served me well for a long time, but I decided to give
vim-plug a try.  Parallel updates, lazy plugin loading, and post-update
hooks are all good reasons to switch.  It helps that it's mostly
otherwise a drop-in replacement.

I added shortcuts to simplify updating (:PU does update + upgrade) and
a block to automatically install vim-plug if it's not already installed.