Add vim mapping for quicker macro replay
Define a normal mode mapping for replaying macro q with just "shift-q".
Macro q is convenient to define (`qq`), and holding shift-q to replay a
bunch of times is much more efficient than typing `@q` repeatedly.  Saw
this tip on Hacker News today, and I already like it and see myself
using it.