Add PlantUML shortcuts for vim to init file and swap arrow directions
Define an 'InitUml' command to write the header and footer directives
for a PlantUML file.  Define a 'UmlArrowSwap' function and leader
mappings to reverse the direction of a PlantUML arrow directive (e.g.
changes "a --> b" to "b <-- a").  Eventually I may make that into a
proper vim plugin, but for now it's just a convenience function that I
source directly.