Add umlnote script
Similar to my mdnote script (which tbh I don't use too often anymore now
that vimwiki is part of my daily workflow), opens an empty plantuml file
with datestamped name in notes directory.

This commit includes a fix to mdnote to open with $EDITOR instead of
insisting on vim.  This cleans up an odd visual issue that appears on
opening as well.