Replace tmux bindings for pane resize
I realized this evening that the bindings I added yesterday prevent
typing the +, _, <, or > characters.  Changed them to
shift-up/down/left/right, which will be much easier to remember anyway.

My previous bindings for shift-left/right were set to swap windows, but
there are already default bindings to { and } for that.

Also fixed up my binding to create a window and immediately prompt for
a name, while I was there.