Add Vim insert-mode mapping to toggle bullets
When working with bulleted lists, I can already toggle between * and -
characters for the lists with <leader>bb in normal mode, but exiting
normal mode for that is a little annoying.  Using <ctrl-b> in insert
mode will do the same toggle.  (Not to be confused with ctrl-b in normal
mode to Scroll Back a page.)