Configure nvm, node, etc. to lazy init
I've found it important to be able to control versions of node closely.
For example, I ran into an issue where webpack's dev server crashed
upon first HTTPS access when using current node (currently v11).  'nvm'
is the preferred way to do that, but adding it to zsh adds a noticeable
delay to shell initialization.

I mostly kept nvm commented out for the last few months, but this week I
found a method to lazy-initialize nvm and node versions it manages.