Switch vimwiki from personal fork to regular release
My personal fork was created when the regular release broke vimwiki for
me.  As long a regular release is working, I prefer it - but if it
gets broken again, I have my fork as a backup.  (I use it every day at
work, so having it working is important to me!)