Remove vim right arrow insert mode mapping
For some reason when I do `inoremap <expr> <Right> expression()`, Vim
ignores the '<expr>' part of the definition and types "expression()"
directly.  It doesn't do this when mapping <CR>, <Tab>, etc.  <Right> is
a "special key", but I didn't see anything in the docs or online
searches indicating this limitation.

Maybe there's some other way to get this to work, but the existing
mapping is driving me nuts, so out it goes. :)