Update tmux pane border style config
In tmux 2.9 the pane-border-* and pane-active-border-* options are
removed in favor of pane-border-style and pane-active-border-style.
Version 2.8 supported both, but the latest 'brew update' causes
errors.  Switching to the new option should fix the errors in all
the tmux versions I currently use (2.9 in Homebrew on Mac, 2.8 in
nix on Linux).