Fix colors for vim in vimrc
I switched to neovim some time ago, and don't know when colors stopped
working correctly for regular vim, but I noticed colors were broken by
something in my vimrc.  This change appears to fix the problems.  Not
that I plan to switch away from Neovim, but better for my configs to be
vim-compatible. :)