Set conceallevel to 0 in vimwiki pages
I've gotten tired of having backticks and asterisks disappear to support
Markdown formatting in my vimwiki pages.  Trying to write a line with
dollar signs in it was the last straw.

The indentLine plugin overrides user preference for conceallevel, so
that plugin is now disabled by default on vimwiki pages.