Add shellrepl script
Use FZF preview to make a "live REPL" for shell command pipelines.
For example, run `cat myfile | shellrepl` to open an FZF prompt, then type
`grep "find" | sed 's/find/replace/g'`.  The results of the expression
will appear in the FZF preview window as you type.

When you press enter, the expression typed is output to the terminal. On
a Mac, the expression will also be copied to the clipboard.