Add timeprompt zsh plugin
Define zsh functions to add the current time to the shell prompt, and
remove it.

In some cases it's useful to have timestamps in a shell session to
record how long a command took to complete - but always having the time
in the prompt would be annoying.  Having an easy set of commands to
enable/disable timestamps is a good compromise.