Switch Vim LSP plugin to coc.nvim
The coc.nvim plugin has become much easier to use than when I first tried
it, and it feels more robust.  It also helps that I was able to configure
it the way I like (e.g. only pop menu on ctrl-space, not continuously while
typing).  It's super nice to be able to rebind my same leader mappings
to the new plugin too - also added new binding for organizing imports,
which I've had in Java within Eclipse for forever.