Set REPORTTIME in timeprompt.zsh
Use the zsh REPORTTIME feature to output the execution time of commands
run when 'timeprompt' mode is enabled.  REPORTTIME is a feature of zsh;
when set to a number, zsh reports the execution time of commands that
run longer than $REPORTTIME seconds (user and system time, not just
wall-clock time).