Update vim ripgrep/fzf command for content searches
The default `:Rg` command from fzf.vim seems to include directory and
filenames, which is not that useful for my <leader>/ binding to search
for file contents.  (If I want to search filenames, I have a <ctrl>-p
binding for that.)  Override the default :Rg command with a different
one from the fzf.vim website that seems to work better.