Configure coc.nvim bookmarks plugin
We'll see if this replaces normal vim marks for me, but I like that
coc-bookmarks a) puts an indicator in the sidebar, b) can jump forward
and back similar to how I jump to next/prev git hunks, and c) can jump
between files through an fzf list, including to annotated marks.
Traditional vim marks are still better for macros and actions.