Add vim mapping to add/update timestamp to line
So far I like the change to append HH:MM timestamp when copying todo
item to journal in my vimwiki daily diary.  However, there are times
when I either add an event to the journal manually or move the todo item
prematurely, making me want to add or update the timestamp.  Added a
function to add/update the timestamp, and updated the copy-todo shortcut
to use the new function.