Add tz script
`tz` is a script to help when working with timezones.  When the `tz`
script is called with a string argument, it converts the value to a
timestamp, then shows the timestamp in local time and UTC.  When `tz` is
called with no argument, it uses current time.

At work I often have trouble converting mentally between local timezone
and UTC; this script is an attempt to help with that. (Arithmetic is
hard for me apparently. :P )