Add vim/filetype.vim to fix vimwiki diary filetype changes
Often when I create an explore-split from the vimwiki diary (using
:Vexplore or :Sexplore commands to create vertical or horizontal split,
respectively), the diary buffer filetype will change from 'vimwiki' to
'conf'.  It seems that filetype autodetection is invoked in this case,
and my default format for diary files (started with my ":InitDiary"
command) looks like the conf filetype to vim.  If a filetype.vim file is
present in the vim user runtime directory, it will run before vim's
built-in filetype detection logic, allowing the default detection to be
overrided in this case.

For this config to work, it must be symlinked into the runtime directory
(~/.config/nvim/filetype.vim).