Update coc.nvim settings for diagnostic messages
Use 'diagnostic.enableMessage = jump' to make the error diagnostic
message only appear when jumping to an error.  (In general I prefer the
editor to show the file content unless I actively perform a command that
changes the view - so a hover command is OK, but automatically popping
up help text while I type is something I'd turn off.  I started to look
into a method for making the diagnostic floating window appear on
keystroke like the hover command, but this option seems good enough for
now.)

Also set 'diagnostic.enableHighlightLineNumber = false'.  The
error/warning icon is enough of a visual cue for me for now.