Replace vim paren autoclose plugin
The vim-autoclose plugin had a conflict with a recent update to coc.nvim
that caused up and down arrows to print out mappings instead of moving
the cursor while in insert mode.  doorboy.vim appears to have similar
behavior (possibly even superior behavior with regard to spaces inside
parens/brackets), but no signs of conflicts with coc.nvim so far.