Add Hammerspoon config to control Bluetooth headphones
I have a set of bluetooth headphones that I really like, but one
annoyance is that it connects with my Mac laptops even when they are put
to sleep.  If I forget to disconnect my headphones from my personal
laptop, it will connect to my home laptop when I'm logging into work,
making me have to go open my home laptop and disconnect them manually.

Hammerspoon is a tool for macOS automation through a Lua API, and
blueutil is a command line utility for controlling Bluetooth devices.
This change registers a Hammerspoon plugin (called a "spoon") that disables
my headphones when my Mac is put to sleep.  It also registers a global
keystroke (ctrl-option-cmd-H) to toggle connect/disconnect to my
headphones, making re-connecting after waking a little quicker.

To use it:
- install Hammerspoon (`brew install hammerspoon`)
- Symlink the Hammerspoon config into home dir
  (`ln -s $DOTFILE_HOME/hammerspoon $HOME/.hammerspoon`)
- install blueutil (`brew install blueutil`)
- create $DOTFILE_LOCAL_HOME/headphonesrc, export a HEADPHONES_MAC env
  variable with the MAC address of the headphones (obtain with
  `blueutil --paired`)