Add login time to vim InitDiary command
I use the InitDiary command to create a basic format for my daily
vimwiki diary entries.  I've started appending timestamps to my daily
activities, and lately one of the first timestamps I (manually) append
is the time when I log in to start the day.  This change adds that
first timestamp automatically.