Add qnotes script
Introduce script to efficiently capture short timestamped messages (e.g.
for vimwiki diary) interactively.  "qnotes" is short for "quick notes".

This is a UX experiment to see if it's faster or better in any way than
my regular vimwiki diary. Idea is to limit ability to edit entries after
they're entered, avoid work to capture timestamps, and even avoid
commands to save changes to the diary file.  The contents of the qnotes
session directory are extremely easy to add into my vimwiki diary
(`cat ~/.qnotes/latest_session/*` outputs the messages), so I don't see
this tool as a replacement for my vimwiki diary.